pyproject.toml
[tool.poetry]
name = "hello-clipseg"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.scripts]
clipseg = "hello_clipseg.main:app"

[tool.poetry.dependencies]
python = "^3.12"
numpy = "^1.26.4"
pillow = "^10.3.0"
opencv-python = "^4.10.0.82"
plotly = "^5.22.0"
torch = "^2.3.1"
torchvision = "^0.18.1"
transformers = "^4.41.2"
typer = {extras = ["all"], version = "^0.12.3"}
pandas = "^2.2.2"
structlog = "^24.2.0"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.2.2"
jupyter = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
